user logout, posts & comments premissions

+++ api/posts/views.py
@@ -1,30 +1,44 @@
 from django.http import Http404
 
+from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status
+from rest_framework.generics import ListCreateAPIView
+from django_filters import rest_framework as filters
 
-from .serializers import PostSerializer, PostLikeSerializer
+from .serializers import PostSerializer, PostLikeSerializer,PostDetailSerializer
 from .models import Post
+from .permissions import IsPostOwnerOrAdmin
 
-class PostListView(APIView):
-    permission_classes = [IsAuthenticated]
+class EventFilter(filters.FilterSet):
+    estate = filters.CharFilter(field_name="estate__name", method='filter_estate')
+    category = filters.CharFilter(field_name="category__name", method='filter_category')
 
-    def get(self, request, format=None):
-        posts = Post.objects.all()
-        serializer = PostSerializer(posts, many=True)
-        return Response(serializer.data)
+    class Meta:
+        model = Post
+        fields = ['estate']
 
-    def post(self, request, format=None):
-        context = {"request": self.request}
-        serializer = PostSerializer(data=request.data, context=context)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    def filter_estate(self, queryset, name, estate):
+        return queryset.filter(estate__name__contains=estate)
+
+    def filter_category(self, queryset, name, category):
+        return queryset.filter(category__name__contains=category)
+
+class PostListView(ListCreateAPIView):
+    serializer_class = PostSerializer
+    filter_backends = [filters.DjangoFilterBackend]
+    filter_class = EventFilter
+
+    def get_queryset(self):
+        return Post.objects.all()
+
+    def perform_create(self, serializer):
+        serializer.save(author=self.request.user)
 
 class PostDetailView(APIView):
+    permission_classes = [IsAuthenticated, IsPostOwnerOrAdmin]
+
     def get_object(self, pk):
         try:
             return Post.objects.get(pk=pk)
@@ -33,12 +47,13 @@ class PostDetailView(APIView):
 
     def get(self, request, pk, format=None):
         post = self.get_object(pk)
-        serializer = PostSerializer(post)
+        serializer = PostDetailSerializer(post)
         return Response(serializer.data)
 
     def put(self, request, pk, format=None):
         post = self.get_object(pk)
-        serializer = PostSerializer(post, data=request.data)
+        self.check_object_permissions(request, post)
+        serializer = PostDetailSerializer(post, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
@@ -46,6 +61,7 @@ class PostDetailView(APIView):
 
     def delete(self, request, pk, format=None):
         post = self.get_object(pk)
+        self.check_object_permissions(request, post)
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
